backend/requirements.txt
anyio==4.12.1
asgiref==3.11.0
astor==0.8.1
attrdict==2.0.1
babel==2.17.0
bce-python-sdk==0.9.59
beautifulsoup4==4.14.3
blinker==1.9.0
cachetools==6.2.4
certifi==2025.11.12
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.1
colorama==0.4.6
contourpy==1.3.2
cryptography==46.0.3
cssselect==1.3.0
cssutils==2.11.1
cycler==0.12.1
Cython==3.2.4
decorator==5.2.1
Django==5.2.9
django-cors-headers==4.4.0
et_xmlfile==2.0.0
exceptiongroup==1.3.1
fire==0.7.1
Flask==3.1.2
flask-babel==4.0.0
flask-cors==6.0.2
fonttools==4.61.1
future==1.0.0
gunicorn==23.0.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
ImageIO==2.37.2
imgaug==0.4.0
itsdangerous==2.2.0
Jinja2==3.1.6
json5==0.9.22
kiwisolver==1.4.9
lazy_loader==0.4
lmdb==1.7.5
lxml==6.0.2
MarkupSafe==3.0.3
matplotlib==3.10.8
more-itertools==10.8.0
mysqlclient==2.2.4
networkx==3.4.2
numpy==1.21.6
opencv-python==4.6.0.66
opencv-python-headless==4.8.1.78
openpyxl==3.1.5
opt-einsum==3.3.0
packaging==25.0
pandas==1.5.3
pdf2docx==0.5.8
pillow==12.1.0
premailer==3.10.0
protobuf==6.33.4
psutil==7.2.1
pyclipper==1.4.0
pycparser==3.0
pycryptodome==3.23.0
PyMuPDF==1.20.2
PyMySQL==1.1.2
pyparsing==3.3.2
pytesseract==0.3.13
python-dateutil==2.9.0.post0
python-docx==1.2.0
pytz==2025.2
PyYAML==6.0.3
RapidFuzz==3.14.3
rarfile==4.2
requests==2.32.5
scikit-image==0.25.2
scipy==1.15.3
shapely==2.1.2
six==1.17.0
soupsieve==2.8.3
sqlparse==0.5.5
termcolor==3.3.0
tifffile==2025.5.10
tqdm==4.67.1
typing_extensions==4.15.0
tzdata==2025.3
urllib3==2.6.3
visualdl==2.5.3
Werkzeug==3.1.5
